+++ numba_dppy/tests/njit_tests/dpnp/test_numpy_linalg.py
@@ -97,7 +97,7 @@ def test_eig(filter_str, eig_input, capfd):
     f = njit(fn)
 
     device = dpctl.SyclDevice(filter_str)
-    with dppy.offload_to_sycl_device(device), dpnp_debug():
+    with dpctl.device_context(device), dpnp_debug():
         actual_val, actual_vec = f(a)
         captured = capfd.readouterr()
         assert "dpnp implementation" in captured.out
@@ -169,7 +169,7 @@ def test_dot(filter_str, dot_name, dot_input, dtype, capfd):
     f = njit(fn)
 
     device = dpctl.SyclDevice(filter_str)
-    with dppy.offload_to_sycl_device(device), dpnp_debug():
+    with dpctl.device_context(device), dpnp_debug():
         actual = f(a, b)
         captured = capfd.readouterr()
         assert "dpnp implementation" in captured.out
@@ -189,7 +189,7 @@ def test_matmul(filter_str, dtype, capfd):
     f = njit(fn)
 
     device = dpctl.SyclDevice(filter_str)
-    with dppy.offload_to_sycl_device(device), dpnp_debug():
+    with dpctl.device_context(device), dpnp_debug():
         actual = f(a, b)
         captured = capfd.readouterr()
         assert "dpnp implementation" in captured.out
@@ -208,7 +208,7 @@ def test_cholesky(filter_str, dtype, capfd):
     f = njit(fn)
 
     device = dpctl.SyclDevice(filter_str)
-    with dppy.offload_to_sycl_device(device), dpnp_debug():
+    with dpctl.device_context(device), dpnp_debug():
         actual = f(a)
         captured = capfd.readouterr()
         assert "dpnp implementation" in captured.out
@@ -244,7 +244,7 @@ def test_det(filter_str, det_input, dtype, capfd):
     f = njit(fn)
 
     device = dpctl.SyclDevice(filter_str)
-    with dppy.offload_to_sycl_device(device), dpnp_debug():
+    with dpctl.device_context(device), dpnp_debug():
         actual = f(a)
         captured = capfd.readouterr()
         assert "dpnp implementation" in captured.out
@@ -269,7 +269,7 @@ def test_multi_dot(filter_str, capfd):
     f = njit(fn)
 
     device = dpctl.SyclDevice(filter_str)
-    with dppy.offload_to_sycl_device(device), dpnp_debug():
+    with dpctl.device_context(device), dpnp_debug():
         actual = f(A, B, C, D)
         captured = capfd.readouterr()
         assert "dpnp implementation" in captured.out
@@ -308,7 +308,7 @@ def test_matrix_power(filter_str, matrix_power_input, power, dtype, capfd):
     f = njit(fn)
 
     device = dpctl.SyclDevice(filter_str)
-    with dppy.offload_to_sycl_device(device), dpnp_debug():
+    with dpctl.device_context(device), dpnp_debug():
         actual = f(a, power)
         captured = capfd.readouterr()
         assert "dpnp implementation" in captured.out
@@ -340,7 +340,7 @@ def test_matrix_rank(filter_str, matrix_rank_input, capfd):
     f = njit(fn)
 
     device = dpctl.SyclDevice(filter_str)
-    with dppy.offload_to_sycl_device(device), dpnp_debug():
+    with dpctl.device_context(device), dpnp_debug():
         actual = f(matrix_rank_input)
         captured = capfd.readouterr()
         assert "dpnp implementation" in captured.out
@@ -359,7 +359,7 @@ def test_eigvals(filter_str, eig_input, capfd):
     f = njit(fn)
 
     device = dpctl.SyclDevice(filter_str)
-    with dppy.offload_to_sycl_device(device), dpnp_debug():
+    with dpctl.device_context(device), dpnp_debug():
         actual_val = f(a)
         captured = capfd.readouterr()
         assert "dpnp implementation" in captured.out