+++ CHANGELOG.md
@@ -4,19 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
 ## [0.17.3] - 2021-11-xx
 
 ### Fixed
+* Enable offloading for `numba.njit` in `dpctl.deveice_context` (#630)
 * Fix upload conditions for main and release branches (#610)
 
 ## [0.17.2] - 2021-11-15
 
-### Changes
+### Changed
 * Use llvm-spirv from bin-llvm during build for Linux and Windows (#626, #627)
 
 ## [0.17.1] - 2021-11-10
 
-### Changes
+### Changed
 * Update clang to icx (#622)
 
 ## [0.17.0] - 2021-11-03